Reworking player model
